// src/app/shared/types/api-types/bidt-absence-dto.interface.ts
export interface BidtAbsenceDTO {
  bidtEmployeeByBidtEmployeeId?: number;
  bidtEmployeeByBidtEmployeeId2Id?: number;
  absenceStartDate?: string;
  absenceEndDate?: string;
  absenceType?: string;
  absenceDescription?: string;
  absenceStatus?: string;
  statusState?: string;
  statusUser?: string;
  statusDate?: Date;
}